requirements.txt
feedparser==6.0.10
requests==2.28.1
langchain==0.3.0
langchain-community==0.3.0
jina==3.34.0
langchain-google-genai==2.0.0
python-dotenv==0.21.0
langchain-chroma>=0.1.2
flask-cors==5.0.0
flask==3.0.0
flask-socketio==5.5.1
redis==4.0.2
sqlalchemy==2.0.16
mysql-connector-python==8.0.28
psycopg2-binary==2.9.3
flask-sqlalchemy==3.1.1